refactor: Restructure Fan.Web folders #294
