+++ src/Fan.Web/Infrastructure/Extensions/IHtmlContentExtension.cs
