+++ src/Fan.Web/Mvc/ViewComponents/CookieConsentViewComponent.cs
@@ -1,11 +1,11 @@
 ﻿using Fan.Settings;
-using Fan.Web.ViewModels;
+using Fan.Web.Mvc.ViewModels;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.Features;
 using Microsoft.AspNetCore.Mvc;
 using System.Threading.Tasks;
 
-namespace Fan.Web.ViewComponents
+namespace Fan.Web.Mvc.ViewComponents
 {
     public class CookieConsentViewComponent : ViewComponent
     {
