+++ src/Fan.Web/Mvc/ViewModels/AnalyticsViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace Fan.Web.ViewModels
+﻿namespace Fan.Web.Mvc.ViewModels
 {
     public class AnalyticsViewModel
     {
