+++ src/Fan.Web/Mvc/ViewModels/HeaderViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using Fan.Membership;
 
-namespace Fan.Web.ViewModels
+namespace Fan.Web.Mvc.ViewModels
 {
     public class HeaderViewModel
     {
