+++ src/Fan.Web/Startup.cs
@@ -4,7 +4,8 @@ using Fan.Blog.Services.Interfaces;
 using Fan.Data;
 using Fan.Membership;
 using Fan.Settings;
-using Fan.Web.Infrastructure;
+using Fan.Web.Infrastructure.Middlewares;
+using Fan.Web.Infrastructure.Theming;
 using MediatR;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
