+++ src/Fan.Web/Themes/Clarity/Views/_ViewImports.cshtml
@@ -5,8 +5,8 @@
 @using Fan.Helpers
 @using Fan.Settings
 @using Fan.Web
-@using Fan.Web.ViewComponents
-@using Fan.Web.ViewModels
+@using Fan.Web.Mvc.ViewComponents
+@using Fan.Web.Mvc.ViewModels
 @using Microsoft.AspNetCore.Identity
 @using System
 @using System.Linq
